Update meal maker
Added 6 total courses into the the menu, 3 for mains and 3 for desserts so a total of 3 appetizers, 3 mains and 3 desserts are on the menu.

Created function getRandomMeal() with no parameters.  Creates 3 variables using getRandomDishFromCourse, passing parameters of each type of course (appetizers, mains, desserts).  

A meal variable is them created with each random course variable assigned as each course type.

A mealPrice variable is given a value of each random course variable price.  

The function then console logs the name of each course and the total price.

added a call to the getRandomMeal() function to generate a random meal with the items and total price console logged